+++ CHANGELOG.md
@@ -13,6 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Defined the entry points of the library using the "exports" field in package.json to make ESM imports more comfortable. This can cause some issues for applications using directory imports from the `lib/build` directory. In those cases we recommend adding `index.js` to the import path.
 
+### Changes
+
+-   `passwordResetPOST`:
+    -   now verifies the email address in all cases if the EmailVerification recipe is initialized
+    -   now tries to link accounts based on account info if AccountLinking is enabled
+
 ## [18.0.0] - 2024-05-23
 
 ### Breaking change